moved formatSchedule to date-utils, use tab-primary
